solution/requirements.txt
asgiref==3.7.2
psycopg2-binary
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==41.0.4
distro==1.7.0
python-dotenv
Django==5.0.1
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
fasteners==0.14.1
future==0.18.2
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
jwt==1.3.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
protobuf==3.12.4
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
pytz==2023.3.post1
pyxdg==0.27
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.5
wadllib==1.3.6
xdg==5
zipp==1.0.0
